Add JLT compatible XML performance output
In addition to the CSV data, which is useful for hand analysis, the perftest
tool now emits an "XML" JTL format file.  This contains the same information
encoded the same way that JMeter does, allowing for interoperability with
additional tools.

This is XML in the sense that it follows the form, but it doesn't actually use
a real XML engine, so output might fail with, eg, embedded newlines, than sort
of thing.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>