Merge performance setup and run into one test set
A limitation of the existing Puppet Acceptance suite is that each test case
file is treated as independent.  A consequence of this is that a failure in
the performance setup ended up with infinitely blocking the performance
simulation run.

This merges both files into one, ensuring that a setup failure prevents a run
of the acceptance tool.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>