Hoist base Slice class into ProjectRazor::Slice
All the common slice functionality was held in the `ProjectRazor::Slice::Base`
class, which was quite confusing when you consider that every implementation
is-a slice, and they are used in compliance with the Liskov substitution
principal by the rest of the system.

This rewrites the code to implement all the functionality in a base class,
with subclasses nested below it.  There is no change to the capability of the
system, just less surprising code structure.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>