Remove "slices loaded" and associated global state
The code around "loaded" slices was a bit of a mess: it actually looked to see
what was available on disk, not just loaded.  It also used global state to
pass data between methods that were separated in time, but not
actually decoupled.

With the rework of condition handling for missing slices we can eliminate most
of that, then inline the "discover available slices" facility into the method
that prints available slices.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>