Move CLI handling into a class, not a script
This moves the command line handling and dispatch out of the `bin/razor`
script and into a class.  There is no current functional change, but rather a
cleaner separation of objects.

This opens the door to better testing of the command line dispatch related
code, to ensure that it behaves correctly, than the script permitted.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>