Implement Puppet Labs best-practice version handling
Instead of a configuration, or source, file containing a hard-coded version
Puppet Labs uses a pair of methods to determine the version of software:

In the case of pre-release software, software installed from git, or nightly
builds we strongly prefer to use the results of `git describe` to identify a
robust and reproducible version number.

In the case of release software we burn the release version string into the
final product, typically by modifying a `lib/${project}/version.rb` file that
defines a `Project::VERSION` constant that third parties can use.

As with our standards, `require "project/version"` doesn't cause the entire
suite to be loaded, so you can use that to make version-specific decisions
about interacting with the rest of the software suite.

Many of those capabilities are not presently used, or usable, with Project
Razor, but following standards ensures we can do that as we export more usable
code-level APIs.

It does improve reporting of the version that happens for current git
installations of the software.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>