Add PuppetLabs packaging automation support to Rake
This adds the `packaging` namespace, which can bootstrap our package building
automation support, as well as the metadata defining where that comes from and
how to update the set of packaging tasks.

This was primarily intended as an internal support infrastructure for building
and releasing the software, but it should be possible to use externally.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>