Use upstream Puppet module artifact in acceptance tests
The acceptance tests for Razor installed the module from our forge; this was a
good simulation of the direct user experience, but didn't match our desired
path: we really wanted to be able to test against the latest module at
all times.

This updates the acceptance suite to work hand-in-hand with our Jenkins
infrastructure to test using a module file built in the upstream
job, automatically.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>