Work around Puppet Module Tool limitations
The Puppet module tool is unable to install a local module that is not present
on the forge, because it depends entirely on remote dependency resolution but
does not send local data to the remote server.

This works around that by forcibly installing the module without dependencies,
then installing every missing module from the remote forge with a script.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>