Update Chef broker to new plugin API and add metadata.
The following metadata items are added to the Chef broker:

* chef_server_url - the URL for the Chef server
* chef_version - the Chef version (used in gem install)
* validation_key - the contents of the validation.pem file
* validation_client_name - the validation client name
* bootstrap_environment - initial Chef environment for node
* install_sh_url - the Omnibus installer URL
* chef_client_path - an alternate path to chef-client binary
* base_run_list - an optional run_list of common base roles

The rendered bootstrap template is no longer written to the client node
as it contains sensitive data (validation.pem contents and future
encrypted data bag secret). Instead the resulting template is logged on
Razor's debug level is debugging is required.