Add a unit tests for the Puppet broker install template
This defines a new - if trivial - unit test file for the Puppet broker agent
install template.  At the moment it validates only that the file compiles, but
as we build up the complexity of the install process it will need to do more
checking to ensure that the generated output is correct.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>