Implement Puppet Labs Debian packaging automation
This adds the infrastructure needed to build nightly Debian packages of
Project Razor, based on the current install model.  This is an initial step
toward the goal of packaged releases of Razor, not a final product:

These are far from "production" quality Debian packages; they install into
/opt/razor to mirror the current git distribution model, are not integrated
into the init system, and otherwise continue to operate as they did before.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>