Support multiple razor install types in acceptance suite
In order to transition smoothly from git to packaged versions of Razor, there
needs to be a period when we support both, so that we don't leave users high
and dry with a broken system.  Even an alpha product like this should care
about user experience at this level.

This enhances the acceptance test installer for Razor to support "package"
install mode, if we ask for `internal-packages` install mode.  The default
remains git, and we can explicitly request that also.

It also configured, in `internal-packages` mode, the internal apt package
repository that contains the artifacts of the last successful internal
package build.

Neither of those is expected to be useful from outside the walls of
Puppet Labs, but the general infrastructure will: when we establish public
package repositories, we will be able to integrate a test mode that uses
those, and ensure that we continue to pass acceptance tests with our packaged
product in the face of external change.

This is part of #307, but not sufficient to complete it.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>