(#366) Broken dependency on `which` in Debian packaging
The Debian packaging indicates a dependency on the package which, which does
not exist. The correct dependency is on the package debianutils which includes
the which(1) command - except that is an "essential" package, so we shouldn't
depend on it, we should just assume it is present.

This drops the dependency in the obvious way, fixing the problem, and
closing #366 along the way.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>