Support the `mongodb-10gen` database as well as `mongodb`
We generally installed using the upstream 10gen packages of MongoDB, even when
there were distribution packages, with the git modules.  As part of smoothing
the migration between the two we should make the `mongodb-10gen` package
satisfy the dependencies of the Razor package.

That should allow both modes - distribution packaged mongodb, or upstream, to
work correctly as the developer wishes.

This is not an issue on the RPM side of the fence, since their policies forbid
any direct dependency between our package and the storage engine.

This closes #368.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>