Puppet dislikes top level brace surrounded variables...
Apparently some versions of Puppet dislike `${foo}` at the top level, leading
to test failures with the message "Could not match ${foo}", which is awesome.

This corrects that - the unquoted variable is portable through all the
variants of the DSL, so solves the problem smoothly.

This is part of #372, but does not resolve it.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>