Razor packages must not overwrite git deployments
The general opinion of the community was that installing packages over the top
of an existing git deployment of Razor is a bad idea, and they should refuse
to install.

I tend to agree: that is the most conservative approach, and puts the
responsibility on the end user to decide how they want to handle the
change over.

This shouldn't last forever, but we should at least initially have pre-install
and pre-upgrade code that refuses to install if /opt/razor/.git exists.

This implements that using pre-install scripts to check for the git directory
being present - but only for the first install; upgrades ignore it, so that
managing configuration through git will work after that first install is
pushed through.

This closes #322.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>