Test packages refuse to install over git deployments
This adds an acceptance test to verify that we don't install packages over the
top of an existing git deployment.  This should help ensure that our
preinstall scripts actually work to abort in that case.

In the words of Knuth, "Beware of bugs in the above code; I have only proved
it correct, not tried it." -- in this case, I have verified it as far as I am
able, but limitations of internal package repositories, multiple Jenkins
deployments, and branch testing mean that I am unfortunately unable to run
this through a full CI suite in the real world.

Consequently I expect something to blow up when faced with reality, although
it should only be a minor deviation in the road.  (None of this is helped, of
course, by bugs in other products that make it harder to verify success or
failure than I might like.)

Signed-off-by: Daniel Pittman <daniel@rimspace.net>