Handle platforms with "purged" packages in git install test
When we validate that packages refuse to install over the top of an existing
git-based deployment, we end up with a package that isn't installed.

Unfortunately, that is represented by two distinct states in Puppet, `absent`
and `purged`, with only a subset of platforms supporting the later state, and
even then, only sometimes reporting it.

This means that the test was broken in the real world: it reported `absent` on
my test bench, but put into production it reported `purged`, and we
incorrectly decided that the package was installed when it actually wasn't.

This makes the obvious update to the check, so that we handle both cases, and
should fix the acceptance test failure after #322 merged.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>