Define confdir in redhat spec
In commit 0fd1f7c, init scripts were added for Razor, and packaging artifacts
were updated to install them. However, as written, the redhat spec file
attempts to install the init script using a macro '%{confdir}' that is not
defined. This causes rpm packaging to fail, because the path to the init script
is invalid.  This commit defines the missing %{confdir} macro as ext/redhat so
that the install succeeds and package build completes. This solution was chosen
over just hardcoding the path 'ext/redhat' because macros are generally more
reusable and maintainable, but in this case it feels like a toss-up since we
use this path only once.