Update internal package repository layout
This updates the acceptance tests to handle the new "consistent snapshot of
internal package repositories" process, in conjunction with our
CI infrastructure.

Essentially, the CI system captures the version of the internal Razor packages
we are testing at the start of the run, and the acceptance suite then uses
that consistent snapshot for the entire run.

This avoids a problem where you could wind up with the package versions
changing between the start of an acceptance run and the end of the same run,
triggered when, for example, we commit multiple pull requests in
quick succession.

This shouldn't have any user-visible change outside our own testing and
package release process.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>