Fix tests after fixing invert vs inverse tag matching
In #294 we fixed a problem where invert vs inverse were used in tag matching,
and they led to inconsistent results as we set one but looked for the other.

Fixing that went well, but a couple of tests were missed, and the result was a
broken system.  This fixes the missed tests, and should clean up the results.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>