(#383) Added postinstall script to kickstart.erb
This change adds the necessary lines at the bottom of the kickstart.erb
template for Red Hat 6 so that the postinstall script is called on
reboot. This is done by downloading the script to the /tmp/postinstall.sh
file, making it executable and adding a line to the rc.local so that
it's eventually called. The code is inspired from other preseed files
for Ubuntu that essentially do the same thing.

The change is necessary because, otherwise, the provisioning process
never even gets to the point of informing the broker. This affected
RHEL 6, CentOS 6 and other distributions based on Red Hat 6.