Update internal package Microkernel ISO location
This migrates the Microkernel ISO image from the older layout to the new,
atomic-snapshot publication model.  That mirrors the same change in the
internal packages, and applies in the same circumstances.

This integrates the acceptance suite with the same testing and promotion
model, which should make it easier to automate publication.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>