Remove leading epoch from spec file changelog
This looks like a lingering artifact from the facter spec file. There was an
epoch of 1 in the spec file previously, and so the changelog had a leading 1 in
the version section. That wouldn't have affected the package but could be
confusing to those reading the changelog. This commit removes that epoch from
the changelog.