(#388) Update to handle three flavours of Microkernel
During the earlier builds I forgot that we build three flavours of
Microkernel, the production, development, and debug types.  The tests and
other CI infrastructure then only tested one, which was not sufficient for
the project.

This fixes up the internal repo layouts to respect the three different types,
and enables testing and releasing all of the different flavours in our nightly
builds area.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>