(#403) Requiring require_all in broker.rb
This change requires require_all in the broker.rb file the same was as
in the model.rb file. This removes the need to explicitly require each
broker under the correspondingly named subdirectory.

The change is desirable because it enables people writing custom brokers
to just drop their Ruby file under the broker subdirectory and have it
required automatically by Razor.