Update `Gemfile` to reflect current gem dependencies
Over time the gem dependencies for Razor have changed, especially with the
introduction of a real unit test suite with a lighter weight dependency chain
than the acceptance suite.

This updates the `Gemfile` to reflect that, allowing us to use that for all
our automated testing needs.  That helps keep this accurate, which hopefully
also makes life easier for new developers on the project.

It also eliminates a number of unused gems, including `autotest`, that had
fairly heavy (and sometimes broken) dependencies.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>