[packaging] Update mocks for rpmbuilder mock format
The Puppet Labs mocks created by rpmbuilder have assumed a new naming format,
pl-el-* vs el-* and pl-fedora-* vs fedora-*, in order to avoid overwriting the
configurations supplied by the mock package. This commit updates the mocks in
Razor to reflect the new standard, so that we can continue to build packages
with builders created with the rpmbuilder module. As we add new builders that
do not have legacy configurations, this will become critical to avoid broken
pipelines.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>