(#417) Add gems over-zealously removed from Gemfile
The Chef cookbook currently depends on the Gemfile for runtime installation of
the dependent gems in Razor; they discovered that a couple of gems that are
still needed, daemons and pg, were eliminated in the last update of this file.

This restores them to their place in the order of things, and should fix the
third party chef cookbook, at least for now, and so closes #417.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>