Remove hard rspec dependency in Rakefile
Razor's Rakefile requires rspec/core/rake_task to add spec tasks to the task
list. This is useful when running tests, but should not be required when
building packages. This commit addresses that by rescuing the LoadError on
rspec/core/rake_task and only adding the spec related tasks if the
Rspec::Core::RakeTask constant is defined. This allows packaging tasks to be
run without rspec installed, but will still allow tests to be run and displayed
if rspec is installed and available.