Citrix XenServer Support for 6.1 (tampa) and 6.0 (boston)
Adds the ability to provision Citrix XenServer.  Model configuration handles
generation of the xenserverinstall_xml answerfile which it uses during installation.

pxelinux.0 is chainloaded to boot the install instead of iPXE due to mboot.32
complications which will be looked at in the future.  It calls back for a generated
isolinux file to obtain it's boot settings for the installation.