Strip the config testing out of the persistence layers test suite
The acceptance tests for the data object still made assertions about the
configuration; this eliminates that, now that we have separated it from the
persistence layer.

All the testing that this performed has also been ported to the regular unit
test suite, so there is no loss in eliminating this change.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>