Eliminate use of config through the persistence layer in tests
The acceptance suite had a bunch of uses of the configuration data through the
persistence layer that, for some crazy reason, didn't show up in my earlier
test runs.

This eliminates those, allowing the tests to actually pass when fully run,
after the initial work disentangling the two concepts in the code.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>