Update tag slice tests to work regardless of IP in configuration file
The tag slice acceptance tests were trusting the content of the Razor
configuration file to locate their next step - even if that was mutable state
changed by other tests.

This updates them to simply assume the Razor instance they are testing is
running on localhost in *all* of their URLs, rather than just the ones that
start the testing process.

This ensures that the tests pass with less need to assume a previous test has
reset configuration on disk to the system defaults.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>