Fix in-memory persistence layer collection handling
The in-memory persistence layer was implemented with some incoherence, and
handled indexing to locate a specific document incorrectly.  It tried to treat
an array of all instances as a hash, rather than finding the record directly
in memory.

As a side effect I was able to delete a few extra lines of code by eliminating
one stateful variable, and using a Ruby feature to fill in blank values in the
target hash appropriately.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>