Fix the server configuration spec for the in-memory persistence update
The configuration object in Razor is stateful.  This means that change
propagates between our tests freely, if they run in the same Ruby VM, and we
end up with the "use in-memory persistence" configuration in the global
helpers causing failures in the config specs.

To combat this the global state is now reset in the configuration code tests,
to ensure that we continue to enter them in the state that we believe we
should be in.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>