(#323) Fix final URL related bug in configuration related acceptance tests
The earlier merge was tested improperly, and wound up full of breakage that
only showed up in the "live" CI system, not on the bench.  While that sucked,
most of the issues were patched.

This one missing URI parsing issue seems to be the last real failure in our CI
system, and with that changed to not double-parse a URI, should give us a
fully functional test suite again.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>