Merge `filtering` module into the one place it is used
The `ProjectRazor::Filtering` module defined a mechanism for general subset
hash comparison, used as part of object retrieval.  This ended up with a
single consumer, the `ProjectRazor::Data` object.

That made it kind of confusing to look into a distinct file when trying to
figure out the implementation - especially because the mixin nature jumped
between the two files for what, in essence, was a single class.

This simple merges the code directly into the data object, and eliminates the
separate file.  That produces the same final result with less files and
lines overall.

It also called out that the `sanitize_hash` method was duplicated between the
utility module - and, so, all classes through the `ProjectRazor::Object` base
class that everything else is derived from - and the exact duplicate
implementation in the filtering module.

As part of the move that was simply dropped; the same code is run, but now
only a single copy exists rather than the two identical copies we had before.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>