Remove a reimplementation of Hash#has_key? from Razor
The `find_key_match` method in the Filtering module - and, now, the data
class - was an open-coded implementation of `Hash#has_key?`.  This eliminates
it in favour of the original method.

It also eliminates the spurious temporary that could only ever have the same
value as the key we were searching on, in favor of an `unless` statement, so
that the actual purpose of the decision point was made clear.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>