Inline the behaviour of `is_valid_json?` in both callsites
The general utility `is_valid_json?` would perform a full JSON parse, and then
discard the results; it treated any error of any sort as a reason to ignore
the problem and carry on.

This was called in two places in a single slice, and can easily enough be
replaced by simply adding a `rescue` block that will ignore all errors.

This cuts our JSON parsing time in half, not that it was terribly significant,
and reduces the complexity of the code overall.

It also makes obvious the somewhat untoward behaviour of that code in terms of
input sanitisation - which while it doesn't fix any problems, at least makes
it clear where they are for future attention.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>