Fix broker plugin listing after script broker was merged
When the script broker was merged it included a new error class, specific to
the broker, that allowed differentiation of remote script errors from other
classes of error.

Unfortunately, that error class was defined at the broker level that Razor
expected to be able to enumerate only broker plugin class children -
triggering a failure in an attempt to list all available broker plugins.

This fixes that by moving the custom error class down a level, to where it no
longer trips up the metaprogramming - and adds a spec test to ensure that this
doesn't explode again.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>