Merge `SliceUtil::Common` module into the one place it is used
The code from `SliceUtil::Common` was actually part of the base `Slice` class,
but was extracted into a separate module.

Despite that, it was extremely intimate with the behaviours of the slice
class, to the point of depending on the exact member variable structure, and
the call sequence of various components.

This merges that code directly into the base class, making it much clearer
that this is not in any meaningful way distinct code from the one consumer.

There should be no functional changes from this, simply a few lines of
boilerplate code removed and a less confusing overall layout of the code.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>