Eliminate `command_shift` method designed only for test stubbing
The `command_shift` method was introduced to improve the ability to stub
behaviour of the Slice base class in tests - but, ultimately, is not actually
needed as the natural behaviour is literally fine for testing.

This removes it, inlining the simple array shift in the two places it
was used.

There is no functional change, since the handful of stubs in the tests were
eliminated during the previous commit.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>