Eliminate several vestigial `command_array` accessors from Slice
There were several definitions of the `command_array` accessor on the `Slice`
object, one through the base object and another defined in the mixin
`SliceUtil::Common` class.

None of them were actually consistently or meaningfully used in the code,
so despite the claim that this might be useful for stubbing, it really had no
practical value.

This commit simply eliminates that code, reducing confusion, API surface area,
and mental bandwidth dealing with this code.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>