Inline the confusingly named `iterate_obj` method on Slice
The Slice base class offered the confusingly named `iterate_obj` method
publicly, and to all children.  This was, in fact, part of the
`add_uri_to_object_hash` method, walking the first layer of values in the map
and transforming any maps in arrays in values by adding the URI to them, also.

This replaces the two call sequence with a single call that applies the
recursive modification, which avoids future developers needing to wonder why
both existed, and what their relationship was.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>