Eliminate unused `get_next_arg` method
The common slice code included a `get_next_arg` method that did a pile of
bookkeeping, but that was never actually used anywhere in the code.

Most of the code either doesn't care about the data that was tracked, or
open-coded some-but-not-all of the same changes; since this is never actually
used, and the data flows are incredibly complex, but the product more or less
does what is expected the shortest path seems to be to eliminate the unused
common function.

There is no behavioural change resulting from this commit.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>