Eliminate unused `throw_uuid_not_allowed_error` helper method
The common slice code had a bunch of helpers that worked in conjunction with
the `get_command_map` abstraction to build a standard set of error handlers
for illegal command line syntax.

The `throw_uuid_not_allowed_error` helper was never used, however, so we can
eliminate it from the codebase.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>