Eliminate 30 unhelpful rechecks that class initialization worked
The slice base class defined a `setup_data` method that it used with wild
abandon, typically at the entry point to each instance method that could be
invoked indirectly in response to the command line.

This was all redundant to the `setup_data` call made in the `initialize`
method, though, which would ensure that the conditional code was never
possible to execute outside the very first creation of the slice.

This eliminates all those `setup_data` calls, resulting in exactly the same
behaviour for the system overall - but substantially less confusion about when
and where the connection to the database is discovered.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>